Simplify the permalinks format to include only the filename
Multiple suggestions mentioned to keep the permalinks as simple as
possible. Unless it is news-based, don't include the date.

That said, there are two options here.
- /:filename/
- /:title/

filename is a better choice to me since it is more "permanent". It
allows me to update the post title without changing the permalink.